DashboarCliente-Service/requirements.txt
flask
flask-cors
pandas
openpyxl
python-dotenv
xlrd>=2.0.1
